# Vendor in required bits of CodedOutputStream and remove protobuf-java…