--- a/exporters/otlp/trace/src/main/java/io/opentelemetry/exporter/otlp/trace/AttributeMarshaler.java
+++ b/exporters/otlp/trace/src/main/java/io/opentelemetry/exporter/otlp/trace/AttributeMarshaler.java
@@ -5,8 +5,6 @@
 
 package io.opentelemetry.exporter.otlp.trace;
 
-import com.google.protobuf.CodedOutputStream;
-import com.google.protobuf.WireFormat;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.proto.common.v1.internal.AnyValue;