--- a/exporters/otlp/trace/src/main/java/io/opentelemetry/exporter/otlp/trace/ResourceMarshaler.java
+++ b/exporters/otlp/trace/src/main/java/io/opentelemetry/exporter/otlp/trace/ResourceMarshaler.java
@@ -32,7 +32,9 @@ final class ResourceMarshaler extends MarshalerWithSize {
   private ResourceMarshaler(AttributeMarshaler[] attributeMarshalers) {
     super(calculateSize(attributeMarshalers));
     ByteArrayOutputStream bos = new ByteArrayOutputStream(getSerializedSize());
-    CodedOutputStream output = CodedOutputStream.newInstance(bos);
+    CodedOutputStream output =
+        CodedOutputStream.newInstance(
+            bos, CodedOutputStream.computePreferredBufferSize(getSerializedSize()));
     try {
       MarshalerUtil.marshalRepeatedMessage(
           Resource.ATTRIBUTES_FIELD_NUMBER, attributeMarshalers, output);