… dependency

Because we only use the version of CodedOutputStream that writes to an OutputStream in both gRPC and okhttp marshalers, this only vendors in that version. We end up with about 1300 loc for the coder - by saying goodbye to the 700KB protobuflite or 1.5MB protobuf libraries, I think this is worth it as binary size will not be a reason to avoid the proto serializer. Checking the class files locally, it's about 23KB of uncompressed code. This is a lot smaller than jackson-core too for reference which would likely be our dependency for a JSON encoding approach. The code is 99.9% copied as-is, just with much deletion, and one line with a boolean check on unsafe removed.

@jsuereth called it that removing the buffering slowed things down (the approach I took was to remove the `AbstractBuffered` and replacing writes to the buffer with `out.write`). So didn't write anything custom, the code is as-is. Will follow up with some ideas for actually making changes to improve the performance, for example threadlocal reused CodedOutputStream and checking the varint encoding.

I removed support for Unsafe to reduce the amount of code vendored and because it performs significantly better without it :O I guess Unsafe is really for the byte buffers, byte arrays are very fast on the JVM already.

After
```
Benchmark                                                                         (numSpans)  Mode  Cnt      Score      Error   Units
MarshalerInputStreamBenchmarks.marshalToByteArray                                         16  avgt   10     13.004 ±    0.346   us/op
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.alloc.rate                          16  avgt   10   1432.026 ±   36.460  MB/sec
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.alloc.rate.norm                     16  avgt   10  29304.005 ±    0.001    B/op
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.churn.G1_Eden_Space                 16  avgt   10   1434.033 ±   78.589  MB/sec
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.churn.G1_Eden_Space.norm            16  avgt   10  29341.737 ± 1246.071    B/op
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.churn.G1_Old_Gen                    16  avgt   10      0.008 ±    0.007  MB/sec
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.churn.G1_Old_Gen.norm               16  avgt   10      0.171 ±    0.147    B/op
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.count                               16  avgt   10    118.000             counts
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.time                                16  avgt   10     50.000                 ms
MarshalerInputStreamBenchmarks.marshalToNettyBuffer                                       16  avgt   10     12.745 ±    0.264   us/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.alloc.rate                        16  avgt   10   1174.283 ±   23.607  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.alloc.rate.norm                   16  avgt   10  23574.688 ±    0.334    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Eden_Space               16  avgt   10   1187.772 ±  126.440  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Eden_Space.norm          16  avgt   10  23833.664 ± 2205.844    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Old_Gen                  16  avgt   10      0.017 ±    0.008  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Old_Gen.norm             16  avgt   10      0.339 ±    0.157    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.count                             16  avgt   10     81.000             counts
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.time                              16  avgt   10     37.000                 ms
```

Before
```
Benchmark                                                                         (numSpans)  Mode  Cnt      Score     Error   Units
MarshalerInputStreamBenchmarks.marshalToByteArray                                         16  avgt   10     15.420 ±   0.251   us/op
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.alloc.rate                          16  avgt   10   1208.283 ±  18.331  MB/sec
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.alloc.rate.norm                     16  avgt   10  29344.006 ±   0.001    B/op
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.churn.G1_Eden_Space                 16  avgt   10   1214.766 ±   2.780  MB/sec
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.churn.G1_Eden_Space.norm            16  avgt   10  29504.403 ± 496.366    B/op
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.churn.G1_Old_Gen                    16  avgt   10      0.032 ±   0.013  MB/sec
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.churn.G1_Old_Gen.norm               16  avgt   10      0.769 ±   0.312    B/op
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.count                               16  avgt   10    100.000            counts
MarshalerInputStreamBenchmarks.marshalToByteArray:·gc.time                                16  avgt   10     43.000                ms
MarshalerInputStreamBenchmarks.marshalToNettyBuffer                                       16  avgt   10     15.537 ±   0.564   us/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.alloc.rate                        16  avgt   10    965.034 ±  34.576  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.alloc.rate.norm                   16  avgt   10  23606.519 ±   0.279    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Eden_Space               16  avgt   10    972.034 ±   2.482  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Eden_Space.norm          16  avgt   10  23789.728 ± 850.649    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Old_Gen                  16  avgt   10      0.015 ±   0.007  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Old_Gen.norm             16  avgt   10      0.368 ±   0.161    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.count                             16  avgt   10     80.000            counts
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.time                              16  avgt   10     32.000                ms
```